Removed unused code, and consolidated tests.
