--- stored_requests/backends/db_fetcher/postgres.go
@@ -1,69 +1,0 @@
-package db_fetcher
-
-import (
-	"bytes"
-	"database/sql"
-	"strconv"
-
-	"github.com/prebid/prebid-server/config"
-	"github.com/prebid/prebid-server/stored_requests"
-)
-
-func NewPostgres(db *sql.DB, queryMaker func(int) (string, error)) (stored_requests.Fetcher, error) {
-	return &dbFetcher{
-		db:         db,
-		queryMaker: queryMaker,
-	}, nil
-}
-
-func NewSql(cfg *config.PostgresConfig) (*sql.DB, error) {
-	db, err := sql.Open("postgres", confToPostgresDSN(cfg))
-	if err != nil {
-		return nil, err
-	}
-
-	if err := db.Ping(); err != nil {
-		return nil, err
-	}
-
-	return db, nil
-}
-
-// confToPostgresDSN converts our app config into a string for the pq driver.
-// For their docs, and the intended behavior of this function, see:  https://godoc.org/github.com/lib/pq
-func confToPostgresDSN(cfg *config.PostgresConfig) string {
-	buffer := bytes.NewBuffer(nil)
-
-	if cfg.Host != "" {
-		buffer.WriteString("host=")
-		buffer.WriteString(cfg.Host)
-		buffer.WriteString(" ")
-	}
-
-	if cfg.Port > 0 {
-		buffer.WriteString("port=")
-		buffer.WriteString(strconv.Itoa(cfg.Port))
-		buffer.WriteString(" ")
-	}
-
-	if cfg.Username != "" {
-		buffer.WriteString("user=")
-		buffer.WriteString(cfg.Username)
-		buffer.WriteString(" ")
-	}
-
-	if cfg.Password != "" {
-		buffer.WriteString("password=")
-		buffer.WriteString(cfg.Password)
-		buffer.WriteString(" ")
-	}
-
-	if cfg.Database != "" {
-		buffer.WriteString("dbname=")
-		buffer.WriteString(cfg.Database)
-		buffer.WriteString(" ")
-	}
-
-	buffer.WriteString("sslmode=disable")
-	return buffer.String()
-}
